requirements.txt
asgiref==3.7.1
Django==4.2.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3